requirements.txt
absl-py==1.1.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.2.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
braceexpand==0.1.7
brotlipy==0.7.0
bypy==1.8.4
cachetools==5.2.0
certifi==2022.12.7
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.0.1
click==8.1.7
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.11.0
dataclasses==0.6
datasets==2.16.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==0.8.3
ExifRead-nocycle==3.0.1
fastjsonschema==2.15.3
filelock==3.9.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.60.0
huggingface-hub==0.20.2
idna==3.4
imageio==2.33.1
imagenetv2-pytorch @ git+https://github.com/modestyachts/ImageNetV2_pytorch@14d4456c39fe7f02a665544dd9fc37c1a5f8b635
img2dataset==1.42.0
importlib-metadata==4.11.4
importlib-resources==5.8.0
iniconfig==2.0.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.4
lazy-loader==0.3
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.4
multiprocess==0.70.15
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.24.2
oauthlib==3.2.0
opencv-python-headless==4.9.0.80
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pluggy==1.0.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.12.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.8
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.0
sentry-sdk==1.39.1
setproctitle==1.3.3
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
supervisor==4.2.4
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminado==0.15.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.1.1
tokenizers==0.15.0
tomli==2.0.1
torch @ http://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp38-cp38-linux_x86_64.whl
torchmetrics==0.9.3
torchtyping==0.1.4
torchvision @ http://download.pytorch.org/whl/cu113/torchvision-0.12.0%2Bcu113-cp38-cp38-linux_x86_64.whl
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
transformers==4.36.2
typeguard==4.1.5
typing-extensions==4.9.0
tzdata==2023.4
urllib3==1.26.14
wandb==0.12.21
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@6b5e8953a80aef5b324104dc0c2e9b8c34d622bd
wcwidth==0.2.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.8.0
